requirements.txt
biopython
regex